test(links): give the percent-encoding tests their own file, and source the rules
The Repowise health gate tripped on tests/test_links.py cohesion: two
more classes appended to a file that already mixes every link concern.
The #1332 end-to-end tests move to tests/test_links_percent_encoded.py.

The design section now cites the reference bundle: Obsidian writes
markdown links percent-encoded (primary source), the resolver side and
Obsidian's encoded character set stay [unverified] with fixtures named,
and backslash/entity decoding (#1353) must run before percent-decoding.

Refs #1332
